backend/requirements.txt
fastapi
uvicorn[standard]
streamlit
pydantic
python-multipart
requests
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy
pymysql
python-dotenv
openai
